minimize graph capture output (#162211)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/162211

Currently `OutputGraphGuardsState` is separated out as a serializable interface for `OutputGraph`, but some of the typing around it is incorrect in dynamo's `guards.py` and `output_graph.py`: more fields are used by code than claimed by `OutputGraphGuardsState`, and it works because either the full `OutputGraph` is passed in or the parts that use those fields are dead when `OutputGraphGuardsState` is passed in.

In this PR we try to further separate the necessary fields of `OutputGraph` that should be retained by a full graph capture mechanism, not just limited to dynamo (as it is currently) but also something like make_fx (in the future). Since these fields do not need to be serialized, the result is an intermediate "common" data structure that is between `OutputGraphGuardsState` and `OutputGraph` in the inheritance hierarchy.

Test Plan:
modified an existing test to have interesting guards, otherwise relying on test coverage

Rollback Plan:

Reviewed By: zhxchen17

Differential Revision: D81718791